01_projects/athletes-etl-sqlserver/python/athletes_etl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "athletes-etl"
version = "0.1.0"
description = "Olympic athletes ETL (CSV -> SQL Server) with audit + rejects"
requires-python = ">=3.9"
dependencies = [
  "pandas",
  "sqlalchemy",
  "pyodbc",
  "pyyaml"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
